src/functions.py : 1. Per mnc2file_sph/src/begrun.c, only compute the SolarMassConst for the elements    present, skiping Li, Be, B
src/snapshot-reader.py :
1. Print out just the gas particles and compare to
   data/gadget2/128part-25mpc/snapshot_016/mnc2file_gas.txt

Comparison of output_gas.txt to mnc2file_gas.txt:
    a) Removed metals and structure column in mnc2file_gas.txt
    #) Sorted both numerically
    #) Compared using diff-with-tol.py
    python diff-with-tol.py output_gas_sort.txt tmp_sort.txt ' ' 0.003,0.003,0.003,0.01,0.01,0.01,0.01,0.01,0.01

Not Quite Idetical, but only 45 lines are different. I think this is due to the fact
that the data was saved as floating point numbers and then I'm taking the log10 of
those numbers. This smells like round off error differences between my C and Python
code so I'm not going to worry too much about it